updating challenges
removing monthly challenges and adding weekly challenges 2

+++ cogs/challenges.py
@@ -1,8 +1,5 @@
 from discord.ext import commands
 import discord
-import re
-
-GITHUB_REGEX = re.compile(r"(https://github.com/[a-zA-Z0-9]+/[a-zA-Z0-9]+)")
 
 
 def setup(bot):
@@ -37,14 +34,18 @@ class ChallengeHandler(commands.Cog):
     @commands.Cog.listener()
     async def on_message(self, message):  # Submitted role.
 
-        if message.channel.id == 680851820587122700:  # weekly
-            if message.author.bot:
+        if message.channel.id in (680851820587122700, 713841306253656064):  # weekly
+            if message.author.bot and message.author.id != self.bot.user.id:
                 return await message.delete()
             
-            submitted = self.bot.guild.get_role(687417501931536478)
-            participant = self.bot.guild.get_role(687417513918857232)
-            submission_channel = self.bot.guild.get_channel(729453161885990924)
-            discussion_channel = self.bot.guild.get_channel(680851838857117770)
+            if message.channel.id == 680851820587122700:  # weekly 1 
+                submitted = self.bot.guild.get_role(687417501931536478)
+                participant = self.bot.guild.get_role(687417513918857232)
+                submission_channel = self.bot.guild.get_channel(729453161885990924)
+            else:  # weekly 2
+                submitted = self.bot.guild.get_role(715676464573317220)
+                participant = self.bot.guild.get_role(715676023387062363)
+                submission_channel = self.bot.guild.get_channel(729453201081761862)
 
             if submitted not in message.author.roles:
                 await message.delete()
@@ -61,43 +62,6 @@ class ChallengeHandler(commands.Cog):
                 embed.set_author(name=str(message.author), icon_url=message.author.avatar_url)
                 embed.set_footer(text=f'#ID: {message.author.id}')
                 await submission_channel.send(embed=embed)
-
-        elif message.channel.id == 713841306253656064:  # monthly 
-            if message.author.bot:
-                return await message.delete()
-            
-            submitted = self.bot.guild.get_role(715676464573317220)
-            participant = self.bot.guild.get_role(715676023387062363)
-            submission_channel = self.bot.guild.get_channel(729453201081761862)
-
-            await message.delete()
-
-            links = GITHUB_REGEX.findall(message.content)
-            if not links:
-                return await message.author.send(f'{message.author.mention} Could not find any valid "Github" url.')
-
-            if len(links) > 1:
-                return await message.author.send(f'{message.author.mention} Please only provide one "Github" url.')
-
-            if len(message.mentions) == 0:
-                return await message.author.send(f"{message.author.mention}, Please make sure to mention your team ("
-                                                 f"yourself included)")
-
-            for member in message.mentions:
-                if participant not in member.roles:
-                    return await message.author.send(f"{member.mention} didn't participated in the challenge")
-                if submitted in member.roles:
-                    return await message.author.send(f"{member.mention} has already submitted")
-
-            for member in message.mentions:
-                await member.add_roles(submitted)
-                await member.remove_roles(participant)
-
-            embed = discord.Embed(description=message.content,
-                                  color=message.guild.me.top_role.color)
-            embed.set_author(name=str(message.author), icon_url=message.author.avatar_url)
-            embed.set_footer(text=f'#ID: {message.author.id}')
-            await submission_channel.send(embed=embed)
 
         elif message.channel.id in [680851798340272141, 713841395965624490]:  # Automatic reaction
             await message.add_reaction("🖐️")
